refactor(handlers): collect all handler files
The contributors shouldn't need to worry about handlers. So I extract
them into separate file.
