+++ src/commands/slash/ping.ts
@@ -1,5 +1,5 @@
 import PING_CONSTANTS from "../../constants/ping";
-import { SlashCommand } from "./slashCommands.handler";
+import { SlashCommand } from "../../handlers/slashCommands.handler";
 
 const slashCommand: SlashCommand = {
   name: "ping",
