+++ src/handlers/clientReadyEvents.handler.ts
@@ -1,5 +1,5 @@
 import { Client, Events } from "discord.js";
-import { importFiles } from "../../utils/filesImport";
+import { importFiles } from "../utils/filesImport";
 
 export interface ClientReadyEvent {
   execute: (client: Client<true>) => void | Promise<void>;
